third_party/WebKit/LayoutTests/printing/pageProperty-with-multicol.html

<!DOCTYPE html>
<script src="../resources/js-test.js"></script>
<style>
    @page { margin:0 77px; }
</style>
<div style="-webkit-columns:3;">
    <br>
    <br>
</div>
<script>
    description("Test that special layout for pageProperty() doesn't crash or cause trouble when the document contains multicol.");
    if (window.internals) {
        shouldBeEqualToString("internals.pageProperty('margin-left', 0)", "77");
    } else {
        testFailed("This test can be run only with window.internals enabled");
    }
</script>
